Moved some Orange Twist specific stuff out of utils
