+++ app/assets/js/src/data/shared/migration/index.ts
@@ -1,0 +1,3 @@
+export { getTaggedDbDump } from '../../../data/shared/migration/getTaggedDbDump';
+export { migrateToLatest } from './migrateToLatest';
+export { updateData } from './updateData';
