+++ app/assets/js/src/data/shared/migration/migrateToLatest.ts
