--- app/assets/js/src/data/shared/updateData/index.ts
@@ -1,1 +1,0 @@
-export { updateData } from './updateData';
