--- app/assets/js/src/utils/indexedDB/migration/index.ts
@@ -1,1 +1,0 @@
-export { migrateToLatest } from './migrateToLatest';
